microservice-product/requirements.txt
Flask
flask_sqlalchemy
kafka-python-ng
python-dotenv